server/requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
dnspython==2.6.1
eventlet==0.33.3
flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.6
greenlet==3.1.1
h11==0.16.0
idna==3.11
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
python-engineio==4.12.3
python-socketio==5.11.2
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
urllib3==2.2.3
werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
